requirements.txt
Django==5.1.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0